requirements.txt
python-telegram-bot[job-queue]==20.7
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

